fix(sell): address all 13 PR #61 review findings
Bugs fixed:
- Image.asset → Image.file in live preview (BUG-1)
- Publish/draft error now shown via SnackBar (BUG-2)
- File-too-large maps to correct l10n key (BUG-3)

CLAUDE.md violations resolved:
- setState removed from CategorySelector → Riverpod subcategoriesProvider (FIX-4)
- Live preview extracted to LivePreviewPanel widget (FIX-5)
- ShippingCarrier/WeightRange enums extracted to shipping_types.dart (FIX-6)
- Cross-feature imports replaced with core/ barrel re-exports (FIX-7)
- ViewModel test split into 4 files + helpers, screen test split into 2 (FIX-8)

High-priority items:
- iOS-only app-settings URL replaced with cross-platform dialog (HIGH-9)
- Hardcoded weight labels replaced with l10n keys (HIGH-10)
- 8 missing error l10n entries added to NL + EN (HIGH-11)
- copyWith nullable fields use Function()? pattern for clearing (HIGH-12)
- Semantics labels added to photo step bottom sheet (HIGH-13)

156 tests passing, 0 analyze issues.
