+++ assets/l10n/en-US.json
@@ -508,6 +508,14 @@
     "stepDetails": "Details",
     "stepQuality": "Quality",
     "stepPublishing": "Publishing...",
-    "stepSuccess": "Published!"
+    "stepSuccess": "Published!",
+    "errorNoPhotos": "Add at least one photo to continue",
+    "errorNoTitle": "Enter a title to continue",
+    "errorNoPrice": "Set a price to continue",
+    "publishError": "Could not publish your listing. Please try again.",
+    "draftError": "Could not save your draft. Please try again.",
+    "fileTooLarge": "This photo exceeds the 15 MB size limit",
+    "errorGeneric": "Something went wrong. Please try again.",
+    "openSettingsHint": "Go to Settings to enable camera access for DeelMarkt"
   }
 }
