+++ assets/l10n/nl-NL.json
@@ -508,6 +508,14 @@
     "stepDetails": "Details",
     "stepQuality": "Kwaliteit",
     "stepPublishing": "Publiceren...",
-    "stepSuccess": "Gepubliceerd!"
+    "stepSuccess": "Gepubliceerd!",
+    "errorNoPhotos": "Voeg minstens een foto toe om door te gaan",
+    "errorNoTitle": "Vul een titel in om door te gaan",
+    "errorNoPrice": "Stel een prijs in om door te gaan",
+    "publishError": "Kon je advertentie niet publiceren. Probeer het opnieuw.",
+    "draftError": "Kon je concept niet opslaan. Probeer het opnieuw.",
+    "fileTooLarge": "Deze foto is groter dan de limiet van 15 MB",
+    "errorGeneric": "Er ging iets mis. Probeer het opnieuw.",
+    "openSettingsHint": "Ga naar Instellingen om cameratoegang in te schakelen voor DeelMarkt"
   }
 }
