+++ lib/features/sell/presentation/widgets/details_step/condition_selector.dart
@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 
 import 'package:deelmarkt/core/design_system/colors.dart';
 import 'package:deelmarkt/core/design_system/spacing.dart';
-import 'package:deelmarkt/features/home/domain/entities/listing_entity.dart';
+import 'package:deelmarkt/core/domain/entities/listing_entity.dart';
 
 /// Condition selector using choice chips.
 ///
