+++ test/features/sell/domain/usecases/create_listing_usecase_test.dart
@@ -1,7 +1,7 @@
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mocktail/mocktail.dart';
 
-import 'package:deelmarkt/features/home/domain/entities/listing_entity.dart';
+import 'package:deelmarkt/core/domain/entities/listing_entity.dart';
 import 'package:deelmarkt/features/sell/domain/entities/listing_creation_state.dart';
 import 'package:deelmarkt/features/sell/domain/repositories/listing_creation_repository.dart';
 import 'package:deelmarkt/features/sell/domain/usecases/create_listing_usecase.dart';
